Use the actual integer types for database model types

+++ signer/src/bitcoin/utxo.rs
@@ -102,9 +102,9 @@ pub struct SbtcRequests {
     /// constructing their next UTXO.
     pub signer_state: SignerBtcState,
     /// The minimum acceptable number of votes for any given request.
-    pub accept_threshold: u32,
+    pub accept_threshold: u16,
     /// The total number of signers.
-    pub num_signers: u32,
+    pub num_signers: u16,
 }
 
 impl SbtcRequests {
@@ -165,7 +165,7 @@ impl SbtcRequests {
     }
 
     fn reject_capacity(&self) -> u32 {
-        self.num_signers.saturating_sub(self.accept_threshold)
+        self.num_signers.saturating_sub(self.accept_threshold) as u32
     }
 
     /// Calculates the minimum fee threshold for servicing a user's
